fixed bugs
fixed issue with gem scaling. Soemtimes, the gem would scale 2x it's
normal size if  a gem would spawn/upgrade on the mouse.

resolved another issue where the gem that was being hovered over may not
have been o the same square that was being hovered over. Therefore, the
player clicked on a gem, but the gem next to it might have been
destroyed.